Moved files and refined colors
