+++ apps/admin-x-activitypub/src/views/Inbox/components/Customizer.tsx
